refactor: fix code review comments - clean up
