+++ ee/packages/amplication-git-pull-service/src/app.module.ts
@@ -2,10 +2,7 @@ import { Module } from "@nestjs/common";
 import { GitPullEventModule } from "./git-pull-event/git-pull-event.module";
 import { HealthModule } from "./health/health.module";
 import { ConfigModule } from "@nestjs/config";
-import { RootWinstonModule } from "./winston/root-winston.module";
 import { AmplicationLoggerModule } from "@amplication/nest-logger-module";
-
-const SERVICE_NAME = "amplication-git-pull-service";
 
 @Module({
   controllers: [],
@@ -13,13 +10,12 @@ const SERVICE_NAME = "amplication-git-pull-service";
     GitPullEventModule,
     HealthModule,
     AmplicationLoggerModule.register({
-      metadata: { service: SERVICE_NAME },
+      metadata: { service: "amplication-git-pull-service" },
     }),
     ConfigModule.forRoot({
       isGlobal: true,
       envFilePath: [".env.local", ".env"],
     }),
-    RootWinstonModule,
   ],
   providers: [],
 })
