+++ ee/packages/amplication-git-pull-service/src/git-pull-event/git-pull-event.types.ts
@@ -1,7 +1,7 @@
 // #region enums
 
 export enum KafkaTopics {
-  GitExternalPush = 'git.external.push.event.0'
+  GitExternalPush = "git.external.push.event.0",
 }
 
 export enum GitFetchTypeEnum {
@@ -17,6 +17,18 @@ export enum GitPullEventStatusEnum {
   Created = "Created",
   Ready = "Ready",
   Deleted = "Deleted",
+}
+
+export enum ErrorMessages {
+  AccessTokenError = "failed to create new access token",
+  CreateNewRecordError = "failed to create a new record in DB",
+  UpdateRecordError = "failed to update a record in DB",
+  PreviousReadyCommitNotFound = "failed to find previous ready commit in DB",
+  RepositoryCloneFailure = "failed to clone a repository",
+  RepositoryPullFailure = "failed to pull a repository",
+  CopyDirFailure = "failed to copy files from srcDir to destDir",
+  CleanDirFailure = "failed to remove non-code files from srcDir",
+  GitHostProviderError = "Invalid source control service",
 }
 
 // #endregion
